Added Feature Comparison and updated the  Final Project file
I added a feature comparison document, to keep track of the feature importances between PCA and the rest of the models.  I also went back add in methods from all team members to my final project in an attempt to incorporate everyone's work.  I also added in the proper Data Observations as well.